feat(agent)!: the tool registry is the agent's
`RegisteredTool`, `ToolSet`, `ToolDispatch`, `dispatch_tool`,
`ToolCatalog` and `ToolLease` move from `rig_core::tool` to
`rig_agent::tool` (`registry`, `catalog`), next to `ToolServer`, their
only driver. rig-core's tool module keeps tool authoring: `Tool`,
`ToolEmbedding`, `ErasedTool`, `DynamicTool`, the portable module,
`ToolContext`, the output and result types. The seam a registry needs
goes public: `LivenessFn`, `DynamicTool::into_parts`,
`PortableDynamicTool::into_parts`. The registry tests move with the
registry, paths only. The facade exposes the registry under `agent`.

Migration: `rig_core::tool::{RegisteredTool, ToolSet, ToolDispatch,
ToolCatalog, ToolLease, dispatch_tool}` → `rig_agent::tool::…` (the
`ToolSet`/`dispatch_tool` paths already resolved there).

Also fixes the two `bus/mod.rs` intra-doc links to `HandlerKey` that
S1 left unqualified (CI's doc job caught them).
