+++ crates/rig-agent/src/tool/catalog/tests.rs
