+++ crates/rig-agent/src/tool/mod.rs
@@ -112,12 +112,15 @@
 //! a terminal provider or telemetry concern; Rig does not reconstruct rich
 
 pub mod builtin;
+pub mod catalog;
+pub mod registry;
 pub mod server;
 
+pub use catalog::{ToolCatalog, ToolLease};
+pub use registry::{RegisteredTool, ToolDispatch, ToolSet, dispatch_tool};
 pub use rig_core::tool::{
-    DynamicTool, ErasedTool, IntoToolOutput, PortableDynamicTool, RegisteredTool, Tool,
-    ToolCatalog, ToolDispatch, ToolEmbedding, ToolErrorKind, ToolExecutionError, ToolLease,
-    ToolOutput, ToolResult, ToolSet, dispatch_tool, tool_definition,
+    DynamicTool, ErasedTool, IntoToolOutput, PortableDynamicTool, Tool, ToolEmbedding,
+    ToolErrorKind, ToolExecutionError, ToolOutput, ToolResult, tool_definition,
 };
 pub use rig_core::tool::{ToolContext, ToolContextError};
